+++ lua/zen-mode/view.lua
@@ -219,7 +219,7 @@ function M.fix_hl(win, normal)
   normal = normal or "Normal"
   vim.cmd("setlocal winhl=NormalFloat:" .. normal .. ",FloatBorder:ZenBorder")
   vim.cmd("setlocal winblend=0")
-  vim.cmd([[setlocal fcs=eob:\ ]])
+  vim.cmd([[setlocal fcs=eob:\ ,fold:\ ,vert:\]])
   -- vim.api.nvim_win_set_option(win, "winhighlight", "NormalFloat:" .. normal)
   -- vim.api.nvim_win_set_option(win, "fcs", "eob: ")
   vim.api.nvim_set_current_win(cwin)