+++ CHANGELOG.md
@@ -1,5 +1,33 @@
 # Changelog
 
+## [1.2.0](https://github.com/folke/zen-mode.nvim/compare/v1.1.1...v1.2.0) (2023-04-17)
+
+
+### Features
+
+* **plugins:** wezterm integration ([#61](https://github.com/folke/zen-mode.nvim/issues/61)) ([6b9c522](https://github.com/folke/zen-mode.nvim/commit/6b9c522b5f74706a46309b83d561ae5cac0f67f5))
+
+## [1.1.1](https://github.com/folke/zen-mode.nvim/compare/v1.1.0...v1.1.1) (2023-03-19)
+
+
+### Bug Fixes
+
+* Make fold and vertical fcs zen ([#85](https://github.com/folke/zen-mode.nvim/issues/85)) ([fc7f1fb](https://github.com/folke/zen-mode.nvim/commit/fc7f1fb40a7d13ea34dd27e645e64c8b431a5269))
+
+## [1.1.0](https://github.com/folke/zen-mode.nvim/compare/v1.0.0...v1.1.0) (2023-02-01)
+
+
+### Features
+
+* alacritty plugin ([#68](https://github.com/folke/zen-mode.nvim/issues/68)) ([48a4269](https://github.com/folke/zen-mode.nvim/commit/48a426953205c5556924f0904c82e23a2c161f72))
+* use transparent bg when Normal hl group not having bg ([#81](https://github.com/folke/zen-mode.nvim/issues/81)) ([00b0e0f](https://github.com/folke/zen-mode.nvim/commit/00b0e0f68d6b4ade6e447462533a5a6a5aa39fcb))
+
+
+### Bug Fixes
+
+* `window.backdrop` option is ignored in `.toggle()` API ([#77](https://github.com/folke/zen-mode.nvim/issues/77)) ([#78](https://github.com/folke/zen-mode.nvim/issues/78)) ([a88f1be](https://github.com/folke/zen-mode.nvim/commit/a88f1be193e904f8c08a6ab4d22e923bd44de7de))
+* **tmux:** can get tmux's option value correctly now ([#71](https://github.com/folke/zen-mode.nvim/issues/71)) ([#72](https://github.com/folke/zen-mode.nvim/issues/72)) ([5d8308e](https://github.com/folke/zen-mode.nvim/commit/5d8308ef39c14ecbd6850b56959094aa932285c6))
+
 ## 1.0.0 (2023-01-04)